pyproject.toml
[project]
name = "ai-trade-insights-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "celery>=5.6.3",
    "fastapi[standard]>=0.135.3",
    "langchain>=1.2.13",
    "langchain-google-genai>=4.2.1",
    "pandas>=3.0.2",
    "passlib[bcrypt]>=1.7.4",
    "pyarrow>=23.0.1",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.4.0",
    "resend>=2.27.0",
    "sqlmodel>=0.0.38",
    "supabase>=2.28.3",
]
